Added getFontColorHex and unit test for it
